requirements.txt
discord.py==2.5.2
psycopg2-binary==2.9.10
asyncio-throttle==1.0.2
python-dotenv==1.0.1
aiohttp==3.10.11
Flask==3.1.1
Flask-CORS==4.0.2
requests==2.32.3
Pillow==11.0.0
aiohttp
discord.py
flask
psutil
pymongo
python-dotenv
topggpy
topggpy
